@@ -23,13 +23,19 @@
 	"prefer-stable": true,
 	"prefer-dist": true,
 	"require": {
-		"php": ">=5.5",
-		"icanboogie/common": "^1.2|^2.0",
-		"icanboogie/storage": "^2.0|^3.0"
+		"php": ">=8.1",
+		"icanboogie/common": "^2.0",
+		"icanboogie/storage": "^6.0"
+	},
+	"require-dev": {
+		"phpunit/phpunit": "^9.5"
 	},
 	"autoload": {
 		"psr-4": {
 			"ICanBoogie\\": "lib"
 		}
+	},
+	"autoload-dev": {
+		"Test\\ICanBoogie\\": "tests"
 	}
 }
